pyproject.toml
[tool.poetry]
name = "diswork"
version = "0.1.0"
description = ""
authors = ["FionaLiu <tomato870208@gmail.com>", "JeterChen <dali1756@gmail.com>", "StephenChen <tom9876555@gmail.com>", "LeoWu <defyto13579@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
mysqlclient = "^2.2.4"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"
google-api-python-client = "1.7.9"
google-auth-httplib2 = "0.0.3"
google-auth-oauthlib = "0.4.0"
oauth2client = "4.1.3"
google-auth = "2.3.3"
django-allauth = {extras = ["socialaccount"], version = "^0.62.1"}
django-storages = "^1.14.3"
boto3 = "^1.34.104"
pillow = "^10.3.0"

channels = {extras = ["daphne"], version = "^4.1.0"}
martor = "^1.6.44"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"